# app/docs/swagger.yaml
definitions:
...
  handlers.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  handlers.listElementRequest:
    properties:
      element:
        description: Element to add to the list (can be string, number, object, etc.)
    type: object
...
info:
  contact: {}
paths:
...
  /documents/lists/push:
    post:
      consumes:
      - application/json
      description: Adds an element to the end of an existing list-type document.
      parameters:
      - description: Key of the list document
        in: query
        name: key
        required: true
        type: string
      - description: 'Column family (default: ''default'')'
        in: query
        name: cf
        type: string
      - description: Element to add
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.listElementRequest'
      - description: 'Write option: sync write to disk'
        in: query
        name: sync
        type: boolean
      - description: 'Write option: disable write-ahead log'
        in: query
        name: disable_wal
        type: boolean
      - description: 'Write option: disable slowdown on write buffer full'
        in: query
        name: no_slowdown
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: Status message
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid request or JSON body
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Document not found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Push element to list
      tags:
      - lists
...
  /documents/lists/shift:
    post:
      consumes:
      - application/json
      description: Removes and returns the first element of a list. Returns an error
        if the list is empty.
      parameters:
      - description: Key of the list document
        in: query
        name: key
        required: true
        type: string
      - description: 'Column family (default: ''default'')'
        in: query
        name: cf
        type: string
      - description: 'Write option: wait for sync'
        in: query
        name: sync
        type: boolean
      - description: 'Write option: disable WAL'
        in: query
        name: disable_wal
        type: boolean
      - description: 'Write option: disable slowdown retries'
        in: query
        name: no_slowdown
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: Removed element from the list
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid input parameters
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Document not found or list is empty
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Shift list (remove first element)
      tags:
      - lists
  /documents/lists/unshift:
    post:
      consumes:
      - application/json
      description: Adds a new element to the beginning of a list document. The element
        can be of any JSON-compatible type.
      parameters:
      - description: Key of the list document
        in: query
        name: key
        required: true
        type: string
      - description: 'Column family (default: ''default'')'
        in: query
        name: cf
        type: string
      - description: Element to insert at the beginning
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.listElementRequest'
      - description: 'Write option: wait for sync'
        in: query
        name: sync
        type: boolean
      - description: 'Write option: disable WAL'
        in: query
        name: disable_wal
        type: boolean
      - description: 'Write option: disable slowdown retries'
        in: query
        name: no_slowdown
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: Insertion successful
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid input or JSON body
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Document not found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Unshift list (add to start)
      tags:
      - lists
...
swagger: "2.0"